Many changes and additions to the compatibility layer.
Specifying the current color has been separated into Color (for varying properties) and CurrentColor (for uniform properties).

GPU_primitives.h is a work in progress.  It contains common geometric figures.
GPU_lighting.h is a start on a compatibility layer for lighting.

A small change is needed to font rendering (again) because of changes in the compatibility layer.