Render: local light group option for materials, blender internal feature from the render branch.
When a material is linked in and has a light group override, this can now use
a local group in the scene file, by replacing the linked light group with a
local group that has the same name. A use case might be controlling the specular
highlight on linked character's eyes per scene.

Patch from render branch by Pablo Vazquez.